backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.3.0
postgrest==0.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
yfinance==0.2.28
requests==2.31.0
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
fpdf2==2.7.6
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
